+++ src/lib/messaging/applier/conflict-detection-backfill.ts
@@ -1,0 +1,94 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { PROVIDER_SUFFIXES } from "./conflict-detection-manifest";
+import type {
+  ConflictRegistry,
+  ConflictRegistryEntry,
+  MessagingConflictProbe,
+  MessagingConflictProbeGatewayDeps,
+  ProbeResult,
+} from "./conflict-detection-types";
+
+/**
+ * Build a tri-state `MessagingConflictProbe` from plain openshell runner deps.
+ *
+ * The liveness result is cached so `sandbox list` is issued at most once per
+ * probe instance. A transient gateway failure returns error instead of absent,
+ * preventing a flaky gateway from being persisted as no providers.
+ */
+export function createMessagingConflictProbe(
+  deps: MessagingConflictProbeGatewayDeps,
+): MessagingConflictProbe {
+  let alive: boolean | null = null;
+  return {
+    providerExists: (name) => {
+      if (alive === null) alive = deps.checkGatewayLiveness();
+      if (!alive) return "error";
+      return deps.providerExists(name) ? "present" : "absent";
+    },
+  };
+}
+
+/**
+ * For entries missing `messagingChannels`, probe OpenShell to infer which
+ * channels the sandbox was onboarded with. Safe to call repeatedly. Probe
+ * errors abort the write for that sandbox so future calls can retry.
+ */
+export function backfillLegacyEntryChannels(
+  entries: readonly ConflictRegistryEntry[],
+  probe: MessagingConflictProbe,
+  updateEntry: (name: string, channels: string[]) => void,
+  providerSuffixes: Record<string, string[]>,
+): void {
+  for (const entry of entries) {
+    if (Array.isArray(entry.messagingChannels)) continue;
+    const discovered: string[] = [];
+    let probeFailed = false;
+    for (const channel of Object.keys(providerSuffixes)) {
+      let channelPresent = false;
+      for (const suffix of providerSuffixes[channel]) {
+        let state: ProbeResult;
+        try {
+          state = probe.providerExists(`${entry.name}${suffix}`);
+        } catch {
+          state = "error";
+        }
+        if (state === "present") {
+          channelPresent = true;
+          break;
+        }
+        if (state === "error") {
+          probeFailed = true;
+          break;
+        }
+      }
+      if (probeFailed) break;
+      if (channelPresent) discovered.push(channel);
+    }
+    if (!probeFailed) {
+      updateEntry(entry.name, discovered);
+    }
+  }
+}
+
+/**
+ * Backfill pre-plan registry entries using built-in manifest provider names.
+ * This infers channel presence only; it must not restore legacy credential
+ * hashes. Remove with the `messagingChannels`/`disabledChannels` fallback once
+ * pre-plan registry rows are no longer supported.
+ */
+export function backfillMessagingChannels(
+  registry: ConflictRegistry,
+  probe: MessagingConflictProbe,
+): void {
+  const { sandboxes } = registry.listSandboxes();
+  backfillLegacyEntryChannels(
+    sandboxes,
+    probe,
+    (name, channels) => {
+      registry.updateSandbox(name, { messagingChannels: channels });
+    },
+    PROVIDER_SUFFIXES,
+  );
+}
